+++ framework/CHANGELOG.md
@@ -16,6 +16,7 @@ Yii Framework 2 Change Log
 - Bug #19589: Fix Error reporting in to the `BaseArrayHelper::getValue()` (lav45)
 - Bug #19316: Fix MysqlMutex with same connection but difference database (kamarton)
 - Bug #19507: Fix eager loading of nested one-to-many relations (spo0okie)
+- Bug #19621: Fix login cookie destroyed after authKey change (MarkoNV)
 - Bug #19546: Reverted #19309 (bizley)
 - Bug #19670: Fix Error null check PHP 8.1 `yii\rbac\DbManager` (samuelexyz)
 - Bug #19520: Fix for TIMESTAMP & ROWVERSION columns in MSSQL insert query (darkdef)